fix(auth): preserve token_epoch across re-enroll + harden signed-token verify
Address PR review findings on the signed-host-tokens change.

Epoch monotonicity (Copilot, security): Register used REPLACE INTO, which reset
token_epoch to 0 on every re-enroll. An operator credential cycle (RotateToken
bumps the epoch) was therefore undone the moment the agent re-enrolled, and a
stolen pre-rotate token became valid again. Switch Register to an upsert (INSERT
... ON DUPLICATE KEY UPDATE) that PRESERVES token_epoch while clearing revocation
and the dead grace-window columns, return the row's current epoch, and mint the
new token at that epoch. Credential cycling is now monotonic across re-enroll.

Snapshot re-enroll handling: replace Snapshot.Forget (which erased the local
entry and briefly re-admitted the old token) with Snapshot.Observe(hostID, epoch),
which records the post-enroll state {epoch, not revoked}. The fresh token is
accepted immediately while any pre-rotate token below the epoch stays rejected
with no staleness window; a never-cycled host (epoch 0) is dropped as before.

Signed-token verify hardening (Copilot): reject a MAC that does not decode to the
HMAC-SHA256 width before hmac.Equal (constant-time only over equal-length inputs),
and decode the MAC with base64 Strict() so a non-canonical final char (2 unused
trailing bits) can no longer be flipped while still verifying. A property test
seed surfaced that token malleability; it is now rejected as malformed.

token_epoch column INT -> BIGINT (CodeRabbit) to match the Go int64 model and
avoid overflow under enough cycles.

Updated the in-flight spec delta with the monotonic-epoch guarantee + a
re-enrollment-preserves-the-token-epoch scenario, and added store/snapshot tests.

Already-addressed findings re-verified, no change needed: refresh epoch-bypass
(RefreshToken re-verifies and rejects stale epochs), fail-closed snapshot load on
both server binaries, agent immediate refresh-window check on startup, token_epoch
+ revoked_at indexes, service.New Secret validation, NewSnapshot nil-source guard.

The failing "Server tests" CI check was a transient Go module-proxy stream error
(golang.org/x/crypto download), not a code failure; re-run triggered.

+++ server/endpoint/migrations/00003_host_token_epoch.sql
@@ -6,8 +6,10 @@
 -- The revocation snapshot reloads `WHERE token_epoch > 0 OR revoked_at IS NOT NULL` every few seconds on every replica. Index BOTH
 -- columns so MySQL can index_merge (union) that OR instead of full-scanning the table: an index on token_epoch alone would not be used
 -- while revoked_at is unindexed. The result set (revoked or cycled hosts) is tiny, so the indexed read stays cheap at fleet scale.
+-- BIGINT (not INT) to match the int64 the runtime model carries (signedtoken.Claims.Epoch, store.TokenStatus, RevocationEntries): a
+-- 32-bit column could overflow under enough operator-driven cycles and silently break credential cycling.
 ALTER TABLE enrollments
-	ADD COLUMN token_epoch INT NOT NULL DEFAULT 0 AFTER host_token_issued_at,
+	ADD COLUMN token_epoch BIGINT NOT NULL DEFAULT 0 AFTER host_token_issued_at,
 	ADD INDEX idx_enrollments_token_epoch (token_epoch),
 	ADD INDEX idx_enrollments_revoked_at (revoked_at);
 -- +goose StatementEnd
